Ability to join groups with moderation - Remake v2

+++ packages/oae-principals/lib/internal/dao.js
@@ -27,6 +27,7 @@ const _ = require('underscore');
 const { Group } = require('oae-principals/lib/model');
 const PrincipalsConfig = require('oae-config').config('oae-principals');
 const { User } = require('oae-principals/lib/model');
+const { PrincipalsConstants } = require('../constants');
 
 const RESTRICTED_FIELDS = ['acceptedTC', 'admin:tenant', 'admin:global', 'deleted'];
 
@@ -263,9 +264,7 @@ const updatePrincipal = function(principalId, profileFields, callback) {
   // Ensure the caller is not trying to set an invalid field
   const validator = new Validator();
   const invalidKeys = _.intersection(RESTRICTED_FIELDS, _.keys(profileFields));
-  validator
-    .check(invalidKeys.length, { code: 400, msg: 'Attempted to update an invalid property' })
-    .max(0);
+  validator.check(invalidKeys.length, { code: 400, msg: 'Attempted to update an invalid property' }).max(0);
   if (validator.hasErrors()) {
     return callback(validator.getFirstError());
   }
@@ -420,22 +419,18 @@ const setAdmin = function(adminType, isAdmin, userId, callback) {
  * @param  {String}     callback.token      The token that can be used to verify the email with
  */
 const getEmailToken = function(userId, callback) {
-  Cassandra.runQuery(
-    'SELECT * FROM "PrincipalsEmailToken" WHERE "principalId" = ?',
-    [userId],
-    (err, rows) => {
-      if (err) {
-        return callback(err);
-      }
-      if (_.isEmpty(rows)) {
-        return callback({ code: 404, msg: 'No email token found for the given user id' });
-      }
-
-      const token = _.first(rows).get('token');
-      const email = _.first(rows).get('email');
-      return callback(null, email, token);
+  Cassandra.runQuery('SELECT * FROM "PrincipalsEmailToken" WHERE "principalId" = ?', [userId], (err, rows) => {
+    if (err) {
+      return callback(err);
     }
-  );
+    if (_.isEmpty(rows)) {
+      return callback({ code: 404, msg: 'No email token found for the given user id' });
+    }
+
+    const token = _.first(rows).get('token');
+    const email = _.first(rows).get('email');
+    return callback(null, email, token);
+  });
 };
 
 /**
@@ -446,11 +441,7 @@ const getEmailToken = function(userId, callback) {
  * @param  {Object}     callback.err        An error that occurred, if any
  */
 const deleteEmailToken = function(userId, callback) {
-  Cassandra.runQuery(
-    'DELETE FROM "PrincipalsEmailToken" WHERE "principalId" = ?',
-    [userId],
-    callback
-  );
+  Cassandra.runQuery('DELETE FROM "PrincipalsEmailToken" WHERE "principalId" = ?', [userId], callback);
 };
 
 /**
@@ -462,24 +453,20 @@ const deleteEmailToken = function(userId, callback) {
  * @param  {Object}     callback.userIdsByEmail     An object keyed by email, whose value are the user ids associated to that email
  */
 const getUserIdsByEmails = function(emails, callback) {
-  Cassandra.runQuery(
-    'SELECT * FROM "PrincipalsByEmail" WHERE "email" IN ?',
-    [emails],
-    (err, rows) => {
-      if (err) {
-        return callback(err);
-      }
-
-      const userIdsByEmail = _.chain(rows)
-        .map(Cassandra.rowToHash)
-        .groupBy('email')
-        .mapObject(principalIdEmailHashes => {
-          return _.pluck(principalIdEmailHashes, 'principalId');
-        })
-        .value();
-      return callback(null, userIdsByEmail);
+  Cassandra.runQuery('SELECT * FROM "PrincipalsByEmail" WHERE "email" IN ?', [emails], (err, rows) => {
+    if (err) {
+      return callback(err);
     }
-  );
+
+    const userIdsByEmail = _.chain(rows)
+      .map(Cassandra.rowToHash)
+      .groupBy('email')
+      .mapObject(principalIdEmailHashes => {
+        return _.pluck(principalIdEmailHashes, 'principalId');
+      })
+      .value();
+    return callback(null, userIdsByEmail);
+  });
 };
 
 /**
@@ -513,9 +500,7 @@ const setEmailAddress = function(user, email, callback) {
 
   // The token is correct, change the email address
   const principalsUpdate = { email };
-  queries.push(
-    Cassandra.constructUpsertCQL('Principals', 'principalId', user.id, principalsUpdate)
-  );
+  queries.push(Cassandra.constructUpsertCQL('Principals', 'principalId', user.id, principalsUpdate));
 
   // Remove the token
   queries.push({
@@ -578,23 +563,16 @@ const iterateAll = function(properties, batchSize, onEach, callback) {
   }
 
   /*!
-     * Handles each batch from the cassandra iterateAll method.
-     *
-     * @see Cassandra#iterateAll
-     */
+   * Handles each batch from the cassandra iterateAll method.
+   *
+   * @see Cassandra#iterateAll
+   */
   const _iterateAllOnEach = function(rows, done) {
     // Convert the rows to a hash and delegate action to the caller onEach method
     return onEach(_.map(rows, Cassandra.rowToHash), done);
   };
 
-  return Cassandra.iterateAll(
-    properties,
-    'Principals',
-    'principalId',
-    { batchSize },
-    _iterateAllOnEach,
-    callback
-  );
+  return Cassandra.iterateAll(properties, 'Principals', 'principalId', { batchSize }, _iterateAllOnEach, callback);
 };
 
 /**
@@ -656,9 +634,7 @@ const _updatePrincipal = function(principalId, profileFields, callback) {
     const queries = [];
 
     // Update the principal record
-    queries.push(
-      Cassandra.constructUpsertCQL('Principals', 'principalId', principalId, profileFields)
-    );
+    queries.push(Cassandra.constructUpsertCQL('Principals', 'principalId', principalId, profileFields));
 
     // If the user's email address needs to be updated, we remove the old one from the mapping
     if (isEmailAddressUpdate) {
@@ -679,13 +655,7 @@ const _updatePrincipal = function(principalId, profileFields, callback) {
       }
 
       // Update the cache, if necessary
-      return OaeUtil.invokeIfNecessary(
-        isUser(principalId),
-        _updateCachedUser,
-        principalId,
-        profileFields,
-        callback
-      );
+      return OaeUtil.invokeIfNecessary(isUser(principalId), _updateCachedUser, principalId, profileFields, callback);
     });
   });
 };
@@ -700,10 +670,7 @@ const _updatePrincipal = function(principalId, profileFields, callback) {
  */
 const _deletePrincipalFields = function(principalId, profileFields, callback) {
   // Remove the specified fields
-  const query = util.format(
-    'DELETE "%s" FROM "Principals" where "principalId" = ?',
-    profileFields.join('", "')
-  );
+  const query = util.format('DELETE "%s" FROM "Principals" where "principalId" = ?', profileFields.join('", "'));
   Cassandra.runQuery(query, [principalId], err => {
     if (err) {
       return callback(err);
@@ -774,28 +741,24 @@ const _isEmailAddressUpdate = function(principalId, profileFields, callback) {
  * @api private
  */
 const _getPrincipalFromCassandra = function(principalId, callback) {
-  Cassandra.runQuery(
-    'SELECT * FROM "Principals" WHERE "principalId" = ?',
-    [principalId],
-    (err, rows) => {
-      if (err) {
-        return callback(err);
-      }
-      if (_.isEmpty(rows)) {
-        return callback({ code: 404, msg: "Couldn't find principal: " + principalId });
-      }
-
-      if (isUser(principalId)) {
-        // Update the cache with the raw cassandra row asynchronously to the response. It is not
-        // necessary for this to complete or be successful before we return to the caller. Note
-        // that it is storing a full cache entry, not updating an existing one which is why this
-        // is safe
-        _updateCachedUser(principalId, Cassandra.rowToHash(rows[0]));
-      }
-
-      return callback(null, _getPrincipalFromRow(rows[0]));
+  Cassandra.runQuery('SELECT * FROM "Principals" WHERE "principalId" = ?', [principalId], (err, rows) => {
+    if (err) {
+      return callback(err);
     }
-  );
+    if (_.isEmpty(rows)) {
+      return callback({ code: 404, msg: "Couldn't find principal: " + principalId });
+    }
+
+    if (isUser(principalId)) {
+      // Update the cache with the raw cassandra row asynchronously to the response. It is not
+      // necessary for this to complete or be successful before we return to the caller. Note
+      // that it is storing a full cache entry, not updating an existing one which is why this
+      // is safe
+      _updateCachedUser(principalId, Cassandra.rowToHash(rows[0]));
+    }
+
+    return callback(null, _getPrincipalFromRow(rows[0]));
+  });
 };
 
 /**
@@ -941,9 +904,7 @@ const _hashToUser = function(hash) {
     largePictureUri: hash.largePictureUri,
     notificationsUnread: OaeUtil.getNumberParam(hash.notificationsUnread),
     notificationsLastRead: OaeUtil.getNumberParam(hash.notificationsLastRead),
-    emailPreference:
-      hash.emailPreference ||
-      PrincipalsConfig.getValue(hash.tenantAlias, 'user', 'emailPreference'),
+    emailPreference: hash.emailPreference || PrincipalsConfig.getValue(hash.tenantAlias, 'user', 'emailPreference'),
     acceptedTC: OaeUtil.getNumberParam(hash.acceptedTC, 0),
     lastModified: OaeUtil.getNumberParam(hash.lastModified)
   });
@@ -994,12 +955,9 @@ const invalidateCachedUsers = function(userIds, callback) {
  * @param  {Object}     callback.err    An error object, if any
  */
 const setLatestVisit = function(user, group, visit, callback) {
-  const q = Cassandra.constructUpsertCQL(
-    'UsersGroupVisits',
-    ['userId', 'groupId'],
-    [user.id, group.id],
-    { latestVisit: visit.getTime().toString() }
-  );
+  const q = Cassandra.constructUpsertCQL('UsersGroupVisits', ['userId', 'groupId'], [user.id, group.id], {
+    latestVisit: visit.getTime().toString()
+  });
 
   return Cassandra.runQuery(q.query, q.parameters, callback);
 };
@@ -1029,26 +987,22 @@ const getVisitedGroups = function(userId, callback) {
  * @api private
  */
 const _getVisitedGroupsFromCassandra = function(userId, callback) {
-  Cassandra.runQuery(
-    'SELECT * FROM "UsersGroupVisits" WHERE "userId" = ?',
-    [userId],
-    (err, rows) => {
-      if (err) {
-        return callback(err);
-      }
-      if (_.isEmpty(rows)) {
-        return callback(null, []);
-      }
-
-      const groups = _.map(rows, row => {
-        const hash = Cassandra.rowToHash(row);
-        hash.latestVisit = OaeUtil.getNumberParam(hash.latestVisit);
-        return hash;
-      });
-
-      return callback(null, groups);
+  Cassandra.runQuery('SELECT * FROM "UsersGroupVisits" WHERE "userId" = ?', [userId], (err, rows) => {
+    if (err) {
+      return callback(err);
     }
-  );
+    if (_.isEmpty(rows)) {
+      return callback(null, []);
+    }
+
+    const groups = _.map(rows, row => {
+      const hash = Cassandra.rowToHash(row);
+      hash.latestVisit = OaeUtil.getNumberParam(hash.latestVisit);
+      return hash;
+    });
+
+    return callback(null, groups);
+  });
 };
 
 /**
@@ -1080,6 +1034,108 @@ const getAllUsersForTenant = function(tenantAlias, callback) {
   });
 };
 
+/**
+ * Create a request
+ *
+ * @param  {String}         principalId                 The princiapl who wants to join the group
+ * @param  {String}         groupId                     The group id
+ * @param  {Function}       callback                    Standard callback function
+ * @param  {Object}         callback.err                An error that occurred, if any
+ */
+const createRequestJoinGroup = function(principalId, groupId, callback) {
+  const lastModified = Date.now().toString();
+
+  Cassandra.runQuery(
+    'INSERT INTO "GroupJoinRequestsByGroup" ("groupId", "principalId", "created_at", "updated_at", "status") VALUES (?, ?, ?, ?, ?)',
+    [groupId, principalId, lastModified, lastModified, PrincipalsConstants.requestStatus.PENDING],
+    err => {
+      if (err) {
+        return callback(err);
+      }
+
+      return callback();
+    }
+  );
+};
+
+/**
+ * Update a request
+ *
+ * @param  {String}         principalId                 The princiapl who wants to join the group
+ * @param  {String}         groupId                     The group id
+ * @param  {Function}       callback                    Standard callback function
+ * @param  {Object}         callback.err                An error that occurred, if any
+ */
+const updateJoinGroupByRequest = function(principalId, groupId, status, callback) {
+  const queries = [];
+  const lastModified = Date.now().toString();
+
+  queries.push({
+    query:
+      'UPDATE "GroupJoinRequestsByGroup" SET "status" = ?, "updated_at" = ? WHERE "groupId" = ? AND "principalId" = ?',
+    parameters: [status, lastModified, groupId, principalId]
+  });
+
+  // Execute the queries
+  Cassandra.runBatchQuery(queries, err => {
+    if (err) {
+      return callback(err);
+    }
+
+    return callback();
+  });
+};
+
+/**
+ * Get all requests related to a group
+ *
+ * @param  {String}         groupId                     The group id
+ * @param  {String}         principalId                 The princiapl who wants to join the group
+ * @param  {Function}       callback                    Standard callback function
+ * @param  {Object}         callback.err                An error that occurred, if any
+ */
+const getJoinGroupRequest = function(groupId, principalId, callback) {
+  Cassandra.runQuery(
+    'SELECT * FROM "GroupJoinRequestsByGroup" WHERE "groupId" = ? AND "principalId" = ?',
+    [groupId, principalId],
+    (err, row) => {
+      if (err) {
+        return callback(err);
+      }
+      if (_.isEmpty(row)) {
+        return callback();
+      }
+
+      const hash = Cassandra.rowToHash(row[0]);
+      return callback(null, hash);
+    }
+  );
+};
+
+/**
+ * Get all requests related to a group
+ *
+ * @param  {String}         groupId                     The group id
+ * @param  {Function}       callback                    Standard callback function
+ * @param  {Object}         callback.err                An error that occurred, if any
+ */
+const getJoinGroupRequests = function(groupId, callback) {
+  Cassandra.runQuery('SELECT * FROM "GroupJoinRequestsByGroup" WHERE "groupId" = ?', [groupId], (err, rows) => {
+    if (err) {
+      return callback(err);
+    }
+    if (_.isEmpty(rows)) {
+      return callback();
+    }
+
+    const users = _.map(rows, row => {
+      return Cassandra.rowToHash(row);
+    });
+
+    return callback(null, users);
+  });
+};
+
 module.exports = {
   createUser,
   createGroup,
@@ -1103,5 +1159,9 @@ module.exports = {
   invalidateCachedUsers,
   setLatestVisit,
   getVisitedGroups,
-  getAllUsersForTenant
+  getAllUsersForTenant,
+  createRequestJoinGroup,
+  updateJoinGroupByRequest,
+  getJoinGroupRequest,
+  getJoinGroupRequests
 };
